fix(parser): support update can be linked with and
