+++ src/parser/oceanbase_parser/parser.py
@@ -247,7 +247,8 @@ def p_update(p):
 
 def p_update_set_list(p):
     r"""update_set_list : update_set
-                        | update_set_list COMMA update_set"""
+                        | update_set_list COMMA update_set
+                        | update_set_list AND update_set"""
     _item_list(p)
 
 
@@ -397,15 +398,10 @@ def p_limit_opt(p):
             p[0] = (p[4], p[2])
 
 
-# TODO: need to display the minus here?
 def p_integer(p):
     r"""integer : INTEGER
-                | MINUS INTEGER
     """
-    if len(p) == 2:
-        p[0] = p[1]
-    else:
-        p[0] = -int(p[2])
+    p[0] = p[1]
 
 
 # non-join query expression
